+++ changelog.md
@@ -2,6 +2,7 @@
 - CO-2508 - Use wawi tracking lists
 - CO-2686 - fix missing customer website
 - CO-2786 - fix missing tab name
+- CO-2764 - Prestashop 8.2 compatibility
 - CO-2773 - add gpsr metadata to product description
 
 ## 2.0.1 _2024-07-05_